Версия 2.1.3 ALPHA-3

Третий ALPHA-билд.

Основные изменения:

- нативная поддержка альтернативного компилируемого анонсера Ocelot;
- поддержка SSL-шифрования у внешнего SMTP-сервера (Яндекс.Почта и т.п.);
- удалена система транслитерации из движка (пользователям следует посоветовать использование сайтов наподобие http://translit.net/ для этой цели);
- исправлена ошибка с прописыванием ретрекера у гостей;
- исправлена ошибка с выводом списка релиз-групп;
- исправления путей к изображениям.

+++ upload/library/includes/cron/jobs/tr_make_snapshot.php
@@ -2,17 +2,21 @@
 
 if (!defined('BB_ROOT')) die(basename(__FILE__));
 
+global $bb_cfg;
+
 DB()->expect_slow_query(600);
 
 //
 // Make tracker snapshot
 //
-define('NEW_BB_BT_TRACKER_SNAP', 'new_tracker_snap');
-define('OLD_BB_BT_TRACKER_SNAP', 'old_tracker_snap');
+if (!$bb_cfg['ocelot']['enabled'])
+{
+	define('NEW_BB_BT_TRACKER_SNAP', 'new_tracker_snap');
+	define('OLD_BB_BT_TRACKER_SNAP', 'old_tracker_snap');
 
-DB()->query("DROP TABLE IF EXISTS ". NEW_BB_BT_TRACKER_SNAP .", ". OLD_BB_BT_TRACKER_SNAP);
-
-DB()->query("CREATE TABLE ". NEW_BB_BT_TRACKER_SNAP ." LIKE ". BB_BT_TRACKER_SNAP);
+	DB()->query("DROP TABLE IF EXISTS " . NEW_BB_BT_TRACKER_SNAP . ", " . OLD_BB_BT_TRACKER_SNAP);
+	DB()->query("CREATE TABLE " . NEW_BB_BT_TRACKER_SNAP . " LIKE " . BB_BT_TRACKER_SNAP);
+}
 
 $per_cycle = 50000;
 $row = DB()->fetch_row("SELECT MIN(topic_id) AS start_id, MAX(topic_id) AS finish_id FROM ". BB_BT_TRACKER);
@@ -25,44 +29,78 @@ while (true)
 	$end_id = $start_id + $per_cycle - 1;
 
 	$val = array();
-	$sql = "
-		SELECT
-			topic_id, SUM(seeder) AS seeders, (COUNT(*) - SUM(seeder)) AS leechers,
-			SUM(speed_up) AS speed_up, SUM(speed_down) AS speed_down
-		FROM ". BB_BT_TRACKER ."
-		WHERE topic_id BETWEEN $start_id AND $end_id
-		GROUP BY topic_id
-	";
+
+	if (!$bb_cfg['ocelot']['enabled'])
+	{
+		$sql = "
+			SELECT
+				topic_id, SUM(seeder) AS seeders, (COUNT(*) - SUM(seeder)) AS leechers,
+				SUM(speed_up) AS speed_up, SUM(speed_down) AS speed_down
+			FROM " . BB_BT_TRACKER . "
+			WHERE topic_id BETWEEN $start_id AND $end_id
+			GROUP BY topic_id
+		";
+	}
+	else
+	{
+		$sql = "
+			SELECT
+				topic_id, SUM(speed_up) AS speed_up, SUM(speed_down) AS speed_down
+			FROM " . BB_BT_TRACKER . "
+			WHERE topic_id BETWEEN $start_id AND $end_id
+			GROUP BY topic_id
+		";
+	}
+
 	foreach (DB()->fetch_rowset($sql) as $row)
 	{
 		$val[] = join(',', $row);
 	}
+
 	if ($val)
 	{
-		DB()->query("
-			REPLACE INTO ". NEW_BB_BT_TRACKER_SNAP ."
-			(topic_id, seeders, leechers, speed_up, speed_down)
-			VALUES(". join('),(', $val) .")
-		");
+		if (!$bb_cfg['ocelot']['enabled'])
+		{
+			DB()->query("
+				REPLACE INTO " . NEW_BB_BT_TRACKER_SNAP . "
+				(topic_id, seeders, leechers, speed_up, speed_down)
+				VALUES(" . join('),(', $val) . ")
+			");
+		}
+		else
+		{
+			DB()->query("
+				INSERT INTO " . BB_BT_TRACKER_SNAP . "
+				(topic_id, speed_up, speed_down)
+				VALUES(". join('),(', $val) .")
+				ON DUPLICATE KEY UPDATE speed_up = VALUES(speed_up), speed_down = VALUES(speed_down)
+			");
+		}
 	}
+
 	if ($end_id > $finish_id)
 	{
 		break;
 	}
+
 	if (!($start_id % ($per_cycle*10)))
 	{
 		sleep(1);
 	}
+
 	$start_id += $per_cycle;
 }
 
-DB()->query("
-	RENAME TABLE
-	". BB_BT_TRACKER_SNAP     ." TO ". OLD_BB_BT_TRACKER_SNAP .",
-	". NEW_BB_BT_TRACKER_SNAP ." TO ". BB_BT_TRACKER_SNAP ."
-");
+if (!$bb_cfg['ocelot']['enabled'])
+{
+	DB()->query("
+		RENAME TABLE
+		". BB_BT_TRACKER_SNAP     ." TO ". OLD_BB_BT_TRACKER_SNAP .",
+		". NEW_BB_BT_TRACKER_SNAP ." TO ". BB_BT_TRACKER_SNAP ."
+	");
 
-DB()->query("DROP TABLE IF EXISTS ". NEW_BB_BT_TRACKER_SNAP .", ". OLD_BB_BT_TRACKER_SNAP);
+	DB()->query("DROP TABLE IF EXISTS ". NEW_BB_BT_TRACKER_SNAP .", ". OLD_BB_BT_TRACKER_SNAP);
+}
 
 //
 // Make dl-list snapshot
